Move transport config back to a separate file.
