+++ extensions/amp-ad-exit/0.1/config.js
@@ -15,7 +15,7 @@
  */
 
 import {FilterType} from './filters/filter';
-import {IFRAME_TRANSPORTS} from '../../amp-analytics/0.1/vendors';
+import {IFRAME_TRANSPORTS} from '../../amp-analytics/0.1/iframe-transport-vendors';
 import {user} from '../../../src/log';
 
 /**
