+++ extensions/amp-ad-exit/0.1/test/test-amp-ad-exit.js
@@ -17,7 +17,7 @@
 import * as sinon from 'sinon';
 import {AmpAdExit} from '../amp-ad-exit';
 import {FilterType} from '../filters/filter';
-import {IFRAME_TRANSPORTS} from '../../../amp-analytics/0.1/vendors';
+import {IFRAME_TRANSPORTS} from '../../../amp-analytics/0.1/iframe-transport-vendors';
 import {installPlatformService} from '../../../../src/service/platform-impl';
 import {installTimerService} from '../../../../src/service/timer-impl';
 import {setParentWindow} from '../../../../src/service';
